+++ esperanto-tools-libs/src/dma/MemcpyH2DAction.cpp
@@ -34,7 +34,7 @@ bool MemcpyH2DAction::update() {
   }
 
   auto currentSize =
-    std::min(std::min(availableBytes, size_ - pos_), ctx_.dmaInfo_.maxElementSize_ * ctx_.dmaInfo_.maxElementCount_);
+    std::min((uint64_t)std::min(availableBytes, size_ - pos_), ctx_.dmaInfo_.maxElementSize_ * ctx_.dmaInfo_.maxElementCount_);
   auto cmaPtr = ctx_.cmaManager_.alloc(currentSize);
 
   // add a copy task to the threadpool
